feat: transition to single source of truth for testIDs
